parte 1 historial compras

+++ MercadoEnvio/WindowsFormsApplication1/ABM Usuario/AltaUsuarioPage.cs
@@ -74,7 +74,7 @@ namespace WindowsFormsApplication1.ABM_Usuario
             if (tipoDeUsuarioComboBox.Text.Equals("Cliente"))
             {
                 int tipoDocumento = Convert.ToInt16(ClienteTipoDocComboBox.Text);
-                Double DNI = Convert.ToDouble(ClienteDNITxt.Text);
+                int DNI = Convert.ToInt32(ClienteDNITxt.Text);
                 Cliente nuevoCliente = new Cliente();
                 /*Encriptacion password*/
                 var mesage = Encoding.UTF8.GetBytes(Password);
