+++ MercadoEnvio/WindowsFormsApplication1/ABM Usuario/ModificarUsuarioPage.cs
@@ -106,7 +106,7 @@ namespace WindowsFormsApplication1.ABM_Usuario
 
             if (tipoDeUsuarioComboBox.Text.Equals("Cliente"))
             {
-                Double DNI = Convert.ToDouble(ClienteDNITxt.Text);
+                int DNI = Convert.ToInt32(ClienteDNITxt.Text);
                 Cliente nuevoCliente = new Cliente();
                 /*Encriptacion password*/
                /* var mesage = Encoding.UTF8.GetBytes(Password);
